+++ web/src/pages/tasks/components/index.ts
@@ -2,10 +2,5 @@
  * Tasks page components barrel export
  */
 
-export { ConfluenceSyncTab } from "./ConfluenceSyncTab";
 export { TaskQueryTab } from "./TaskQueryTab";
 export { ScheduledTasksTab } from "./ScheduledTasksTab";
-export { ConnectionCard } from "./ConnectionCard";
-export { BindingTable } from "./BindingTable";
-export { SchedulerStatus } from "./SchedulerStatus";
-export { PageManageDialog } from "./PageManageDialog";